Fix clipboard on plain HTTP: extract safeCopy utility
navigator.clipboard.writeText throws NotAllowedError on non-HTTPS
origins (LAN access via 192.168.x.x). Extract a safeCopy() helper in
lib/clipboard.ts that tries the modern API first and falls back to
document.execCommand('copy') which works on any origin.

Apply to all three places: ShareButton (lobby copy code + share),
ShareResultButton, and results page inline share handler.
